Moved log endpoint to router
